Compatibility fixes for building a 64-bit Windows executable
On 64-bit windows sizeof(long) is 4 and sizeof(void*) is 8 so
code that casts pointers to long or vice-versa needs to be fixed.

In most cases we can replace long with intptr_t.

That fix won't work for wxWidgets GetExtraLong/SetExtraLong and
a pointer cannot be passed through those methods. Instead we can
pass the pointer in pieces in the int and long message fields.

+++ src/myframe.h
@@ -79,7 +79,7 @@ struct VarDelayCfg
     int longDelay;
 };
 
-typedef void alert_fn(long);
+typedef void alert_fn(intptr_t);
 
 class MyFrameConfigDialogPane : public ConfigDialogPane
 {
@@ -444,9 +444,9 @@ public:
     double GetCameraPixelScale() const;
 
     void Alert(const wxString& msg, int flags = wxICON_EXCLAMATION);
-    void Alert(const wxString& msg, alert_fn *DontShowFn, const wxString& buttonLabel, alert_fn *SpecialFn, long arg,
+    void Alert(const wxString& msg, alert_fn *DontShowFn, const wxString& buttonLabel, alert_fn *SpecialFn, intptr_t arg,
                bool showHelpButton = false, int flags = wxICON_EXCLAMATION);
-    void SuppressableAlert(const wxString& configPropKey, const wxString& msg, alert_fn *dontShowFn, long arg,
+    void SuppressableAlert(const wxString& configPropKey, const wxString& msg, alert_fn *dontShowFn, intptr_t arg,
                            bool showHelpButton = false, int flags = wxICON_EXCLAMATION);
     void ClearAlert();
     void StatusMsg(const wxString& text);
@@ -499,7 +499,7 @@ private:
 
     alert_fn *m_alertDontShowFn;
     alert_fn *m_alertSpecialFn;
-    long m_alertFnArg;
+    intptr_t m_alertFnArg;
 
     std::vector<time_t> m_cameraReconnectAttempts; // for rate-limiting camera reconnect attempts
